engine: Upgrade HA VM lease failure during Launch
When upgrading the version from 4.1 to 4.2, HA VMs
with lease SD IDs fail to launch due to a modification
where the lease info was moved from VM Static to VM Dynamic
DB tables. This causes the lease info not to be found and so
the HA VM fails to launch.

This fix removes the lease info validation and
in the run execute detects the issue where the lease
SD ID exists and the lease info is missing in order
to recreate the lease info in order to avoid the
failure.

Change-Id: I1178359c342318b86e3d69694a514868e77fa1cc
Bug-Url: https://bugzilla.redhat.com/1716951
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>