engine: Reset the USB controller for chipset delta
When the USB is enabled and the Bios Type is changed
from Q35 to Default or visa versa, the controllers
do not update because the slots have not changed.

In order to ensure that the controllers are correct,
when the Chip Sets are different and the usb policy
is enabled, the usb spice controllers shall be
removed and readded so that the controllers are
correct. When the Chip Sets are the same, the
update occurs as it did before only if the slots
changed.

Change-Id: I6bc91225d0f53b3ad7fd2df2d6ac1a09640044d5
Bug-Url: https://bugzilla.redhat.com/1715393
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>