engine: Numa config CPUs needs to include all CPUs
The domainXML sent to libvirt when the VM is launched
was sending in the NUMA Configuration section the list
of current CPUs. This led to a libvirt QEMU warning that
this method has been deprecated and now the NUMA
Configuration section should include all CPUs from the
minimum to the maximum as specified within the VCPU section.

Now the NUMA Configuration section shall include the range of
all Virtual CPUs from 0 to the maximum VCPU as sent in the
VCPU section.

Change-Id: Iefb728848bc28b43048d3bd194072730941f5b5a
Bug-Url: https://bugzilla.redhat.com/1437559
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>