core: import vm - update disks size by qemu-img info
Disk's size is stored in OVF with GiB units, hence, the size is
rounded down to nearest GiB value when saved in the OVF.
I.e. if the disk size is not a multiple of a GiB, the saved size
in the OVF will be smaller - which is misleading and problematic
once imported back to the system.

The suggested fix is as follows:
* Import flow: query the 'virtual size' of each image using
               'qemu-img info' and update the db.
* Export flow: round the size value up (instead of down).

Change-Id: I810a37fb036bceb1b24357ad1328daf40291cb3c
Bug-Url: https://bugzilla.redhat.com/1711691
Signed-off-by: Daniel Erez <derez@redhat.com>