engine: add devices to the list of changed fields
In the VMs table, a list of all changed fields
between current and next run configuration is shown in the tooltip of
VM type column. However, the list has contained only changed VmBase fields.

This PR adds also devices to the tooltip. In order to get the list of
changed devices, an instance of VmManagementParametersBase needs to be
created. The creation is similar to that in CloneVmCommand, so the
common parts were extracted to VmDeviceUtils.

Change-Id: I3e875e1f1943cda882e02b3d89a3ee84d343de2a
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>