core: Call disconnectStorage while detaching a master SD
While detaching a master storage domain, there's an execution of
mounting the SD.
Currently, there's no call to "disconnectStorage", so the mount point
remains mounted after the detach operation finished.
By adding a call to "disconnectHostsInUpToDomainStorageServer()", that
storage domain is being disconnected and unmounted.

Furthermore, deactivating the SPM's last domain makes a second call to
"disconnectStorage", which causes a "mountpoint not found" error.
A validation was added in order to avoid this issue.

Change-Id: I3e5bf8fc23038f71bf03538cd44739154de0d474
Bug-Url: https://bugzilla.redhat.com/1494531
Signed-off-by: Shani Leviim <sleviim@redhat.com>