core: add a callback to DestroyImage command
In case of a failure to remove a volume during merge flows (live/cold),
a retry mechanism takes place in case of specific errors and retries
to run the command.

The mechanism will work only if the call for starting the volume removal
on the VDSM wasn't sent successfully.

This mechanism seems to be broken, if the described failure occurs and a
task doesn't create/sent to the VDSM, the CoCo mechanism used in the
parent command endlessly waits for failed DestroyImage command to
finish.

This patch fixes the issue by adding a callback to DestroyImage command
and allow CoCo progress in case of a failure.

Change-Id: Ia6c5dadf3ebad152e14883b0c9208e4710a66833
Bug-Url: https://bugzilla.redhat.com/1637172
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>
(cherry picked from commit db8f54618b807b39d0b125a8b14cf04bf3116d9a)