engine: Build Numa Node CPU ID List as range
When running a VM the domain XML is sent to the libvirt
which includes the Numa Node section containing a list
of CPU IDs. Previously we only sent the online CPU IDs,
but we modified this to send the online and offline CPU IDS.

Because the size of each CPU ID List for each Numa Node can
become quite large, the changes herein allow for building
the CPU ID lists with ranges in order to reduce the size.
Libvirt currently supports both lists containing individual
CPU IDs and CPU ID ranges separated by commas.

Change-Id: I83f3e6c1d22535ee2425fce586d1c293d36596a0
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>