engine: enumerate cluster-wide sync-host events
When syncing all out-of-sync hosts in a cluster an event is created for
each host sync operation start and each host sync operation end.

These events will now be numbered as 'i out of N' in the format already
used for single host sync - 'i/N'.

For a single host sync operation, the numbering will remain 1/1.

Change-Id: I539c263dc80b8ccd77e693703dc2a82fda700b3e
Bug-Url: https://bugzilla.redhat.com/1720994
Signed-off-by: Eitan Raviv <eraviv@redhat.com>