core: decouple refresh vds from update monitoring needed
This patch decouples the update of the cached vds operation from
updating the state of the isMonitoringNeeded flag in VdsManager.
It provides better flow control and visibility for the
isMonitoringNeeded flag update which is now concentrated in
VdsManager#refreshImpl and leaves just a single responsibility to the
VdsManager#refreshCachedVds method.

Change-Id: I5e0a7a234ec9f5f52fa99002067f8815125ea70c
Signed-off-by: Eitan Raviv <eraviv@redhat.com>