Support Event-Notifier API
This patch introduces support for fetching, adding and deleting
subscriptions for the Event Notifier. Event-subscriptions exist
in the context of a user. The ID of an event-subscription is the
event name, which together with the user, uniquely identify the
subscription:

List subscriptions for user xxx:
  GET .../api/user/xxx/eventsubscriptions

Get user xxx's subscription to vm_paused event:
  GET .../api/user/xxx/eventsubscriptions/vm_paused

Remove user xxx's subscription to vm_paused event:
  DELETE .../api/user/xxx/eventsubscriptions/vm_paused

Subscribe user xxx to vm_paused event:
  POST .../api/user/xxx/eventsubscriptions
     <event_subscription>
        <event>vm_paused</event>
     </event_subscription>

* Currently only SMTP (email) notifications are supported via the API
* This patch depends on ovirt-engine-api-model patch:
  https://gerrit.ovirt.org/101093

Bug-url: http://bugzilla.redhat.com/1700338
Change-Id: I800dea0915df4b59414dfc13b1f39c3927540596
Signed-off-by: Ori_Liel <oliel@redhat.com>