core: Add TSC frequency requirement for HP VMs
This patch moves the functionality previously performed by VDSM hook to
the engine, and extends it.

The cpuflags hook added '+invtsc' flag to the CPU. In addition, CPU
TSC timer needs to be added, and TSC frequency is readilly available in
the engine. Therefore the functionality is moved to the engine.

Change-Id: Ia941fc6591938ec3178f5a02585b4ec645951bbd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1679122
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>