webadmin: restrict critical space action blocker value to be valid
In "Storage Domains" page, select a storage domain
and click on "Manage Domain", then "show parameters".

In this change "Critical Space Action Blocker" value
will be validated to be in between 0 and total storage
domain size. If the value is not in range the
textbox will show a warning with a tooltip message
"Critical space action blocker should not be more
than the maximum size of the storage domain ({0}GB)."

Before the change, the value can be any integer
in the range 0 to max integer type value.

Screenshots: https://imgur.com/a/ceo6wSL

This patch fixes the NullPointerException after
the following revert to the patch:
https://gerrit.ovirt.org/c/101144/

Change-Id: I1326d55cf2fa17cb048780bdf7aa60f4a6bb4f57
Bug-Url: https://bugzilla.redhat.com/1533160
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>