improve vm-analyzer
Introduce ObjectIdentityChecker#isAnyFieldChanged that returns true when
the first property that has changed is found, without iterating all properties.

It is now used by VmAnalyzer that now stores the fields that VDSM *may* change
(rather than the fields that are unchangeable by VDSM) as a Set to improve
checking whether a concrete field is contained in that collection.

This way, we no longer need to remove the applications-list when the status
of the VM is not up but rather set the reported list to the existing list,
i.e., to reflect that the reported applications-list should be ignored at
this point. We also don't need to remove the unchangeable fields from the
collection returned by getChangedFields every time the latter is called by
VmAnalyzer.

Change-Id: I09d1abc7d8464d5eccdc2b94a093b8fd1e6d1253
Signed-off-by: Arik Hadas <ahadas@redhat.com>