engine: remove and maintenance lock conflict
The intention is that after the host is in
'maintenance' state, we can remove it.

Before the bug https://bugzilla.redhat.com/1667723
the remove command executed and force removed
the host, because of locking/state issue.
The fix prevented the host from being removed
before it is in maintenance state.

In this bug the engine logs shows that the host
is in 'maintenance' state, but when trying to
remove it, the lock introduced in the previous
bug blocked the action, and the logs shows
that the host did not finish disconnecting
form storage pool servers with the commands
DisconnectHostFromStoragePoolServersCommand and
DisconnectStorageServerVDSCommand.

This commit adds lock with a timeout of
1 minute for removeVdsCommand.

Now after the remove command issued, after
deactivating the host, the command will wait
until disconnect from storage server command
is finished and start removing the host.

Change-Id: Ie0e52af51df8e70e5f9ceefb0d8d9ca5f85b61a9
Bug-Url: https://bugzilla.redhat.com/1720908
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>