webadmin: use vm popup dialog also for clone
When cloning VM, the user could change only the VM name, all other fields
remained the same.

However, there are cases when the user wants to change e.g. the storage
domain of the attached disks. This PR uses the existing model
and widget for creating and editing VM also for cloning the VM. This
allows to modify almost all VM fields during the clone including disks
and network interfaces.

Change-Id: Ife7097fb47964a5f5e77925188d77dceb0f3b946
Bug-Url: https://bugzilla.redhat.com/1696245
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>