webadmin: align status progress bar horizontally with text
After moving a disk from a storage domain
to another, the status progress bar was appended
to the next line which causes the table to misalign
by hiding some rows and shows a scrollbar for that.

Now, the progress bar appended next to the status
text, without affecting the whole table.

screenshots: https://imgur.com/a/4DGrOXr

Change-Id: Ifebe45dfd0975c6e84205c05de68314b5307e79b
Bug-Url: https://bugzilla.redhat.com/1690155
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>