webadmin: Add checkbox to control TSC Frequency
High Performance VMs use TSC Frequency by default to perform safe
migrations. As this is a new feature, users should be able to override
(disable) this functionality for cases where it prevents some unforseen
workflows.

This patch adds a checkbox to migration options. On by default for HP
VMs, it can be unchecked.

Screnshot: https://imgur.com/a/4YrNFHg

Change-Id: If0b4eb82e5e3ad75e90394acebb260e988d118b3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1679122
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>