engine: make vds use cluster cpu_flags
With the introduction of the default/secure CPU concept we need to make
sure that the hosts are not non operational after the configuration of
secure CPU is changed.

This is done by storing the cpu_flags and cpu_verb in the cluster
table. The VDS uses these values for the CPU flags check so it will stay
operational. The flags and verb in the cluster table are set to configured
values automatically in Edit cluster popup view.

Change-Id: Iab49f7486220e01e41caa2016765a6f5b7f1624a
Bug-Url: https://bugzilla.redhat.com/1731395
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>