core: NPE in FeatureSupported in tests
IConfigUtilsInterface.getValue() implementations are expected to throw
IllegalArgumentException when the value is not found. But in tests, when
IConfigUtilsInterface is mocked, null is returned for unknown values,
causing NullPointerException in FeatureSupported.supportedInConfig(),
when Boolean is cast to boolean.

The consistency is especially important in the code that catches
IllegalArgumentException. This code will work in a regular run, but fail
in tests.

Change-Id: If27047ca0bee9513bfd9d757b15cb158a9996308
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>