engine: cluster out of sync indication restructuring
The query for getting the out-of-sync host list to display in the
cluster list view was causing a flicker in the cluster logical networks
view due to too much garbage collection in the browser javascript engine.

This restructuring minimizes the payload and moves all of the
calculation to the back-end to prevent unnecessary object creation in
the front-end. This is accomplished  by passing a simple string with the
host names instead of a collection of VDS objects.

Change-Id: If626c495d4fb49ffac8e3934b6ecd84bf1154fdb
Bug-Url: https://bugzilla.redhat.com/1613702
Signed-off-by: Eitan Raviv <eraviv@redhat.com>