core: Improvements in AutoStartVmsRunner
This patch simplifies concurrent usage of the AutoStartVmsRunner
class. VMs that need to be started are only added to a set.
Then the periodic method takes this set and updates other internal
structures.

This change also removes a race condition, when a VM is manually
started and then crashes while AutoStartVmsRunner is waiting for
the time to start the VM. Previously, the restart counters would
not be reset.

Change-Id: Ib256cfcd0af4c9bc40617bf5c96f727f7b412b25
Bug-Url: https://bugzilla.redhat.com/1325468
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>