core: introduce RemoveUnregisteredVmTemplateCommand
Until now, unregistered templates could not be removed.
The only option was to register the template to the
data center and then remove it.

This patch adds the option to remove an unregistered template.
In order to remove an unregistered template the following call
should be sent -

DELETE
.../ovirt-engine/api/storagedomains/<domain-id>/templates/<template-id>

If a wrong template ID was sent or the template doesn't exist a 404
response will be sent back.

Change-Id: If2b0f0d9b59ab5709b6c6f2177e6972869f68c29
Bug-Url: https://bugzilla.redhat.com/1733932
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>