packaging: setup: pki: Use 3DES to encrypt pkcs12
openssl uses 40 bit RC2 for encrypting the PKCS#12 archive.
This makes it non-fips compliant, so in fips mode, uses instead 3DES.
This means that a p12 file generated on non-fips machine is unreadable
on a fips machine. This is, despite the fact that the encryption/signing
algorithms used for the key and cert included in the p12 _are_ fips-
compliant, by default. So:

Make new generated p12 files use 3DES.

This will make it easier to upgrade new setups from non-fips to fips.
It won't help, in itself, for upgrading existing systems.

Change-Id: Ic42994e4be9c53e623361555f21f1dd58387e44d
Signed-off-by: Yedidyah Bar David <didi@redhat.com>