core: introduce RemoveUnregisteredVmCommand
Until now, unregistered VMs could not be removed.
The only option was to register the VM to the
data center and then remove it.

This patch adds the option to remove unregistered VMs.
In order to remove an unregistered VM the following call
should be sent -

DELETE
.../ovirt-engine/api/storagedomains/<domain-id>/vms/<vm-id>

If a wrong VM ID was sent or the VM doesn't exist a 404
response will be sent back.

Change-Id: I2f00480217eea8916486e3dc3e55b6c8ea14c946
Bug-Url: https://bugzilla.redhat.com/1733932
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>