core: implement removing unregistered disk
Until now, unregistered disks could not be removed.
The only option was to register the disk to the
data center and then remove it.

This patch adds the option to remove unregistered disks.
In order to remove an unregistered disk the following call
should be sent -

DELETE
.../ovirt-engine/api/datacenters/<dc-id>/storagedomains/<domain-id>/disks/<disk-id>

If a wrong disk ID was sent or the disk doesn't exist a 404
response will be sent back.

Change-Id: I566ec8ef5f7e5d2f7ae0cf86261885cab92a342e
Bug-Url: https://bugzilla.redhat.com/1733932
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>