engine: Replace floppy support for Versions >= 4.4
The sysprep floppy driver for windows is being replaced.

For versions 4.4 and greater we shall replace the floppy
that contains the sysprep functionality with an additional
CD ROM. We are still supporting the floppy driver for
versions less than 4.4 such as when a payload is specified
or when a floppy path is specified.

Change-Id: I015bf79cc544da2f042811c70f96ff4b0ace2e8b
Bug-Url: https://bugzilla.redhat.com/1306586
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>