restapi: Enable searching Events for non-admin users
Currently we enable filtering events for non-admin users
or searching events as admin users.

This patch enables the combination: request for events as a
non-admin user (meaning 'filtered') with a search constraint.

Similary to other such cases in the API, this is done by
performing an intersection 'behind the scenes' between the
result of the filtered query, which considers the user's
permissions, and the search query, which disregards permissions.

Bug-Url: http://bugzilla.redhat.com/1731121
Change-Id: I29ecdf09f486a1206cca825160425d24f2d41db8
Signed-off-by: Ori_Liel <oliel@redhat.com>