core: Consider guest tools located in data domain for mark
Added into the consideration any guest tools ISOs located
in Data domains. Changed the comparison of guest tools versions
from String to Version object.

With this fix, the mark for "new guest tools are available"
will consider guest tools in Data domains, not only those in ISO
domain. A new periodically check will be execute to check the version.
The change of comparison to Version object and regex fixes
are fixing the cases for non-one digit version number.

Change-Id: Iaa31b8fe0c1182d7ff5602103f2f7c8468e910a5
Bug-Url: https://bugzilla.redhat.com/1720795
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>