webadmin: add possibility to update tooltip text
ElementTooltipUtils.setTooltipOnElement() always created a new tooltip
even if there already had been one. This caused problems with tooltipped
icons on detail pages when navigating with "Up" and "Down" keys. The
tooltip was OK for the first entity detail, but for the others only icon
was shown and no tooltip.

Now if there already is a tooltip for the element, it is not destroyed
and re-created but only the text is updated.

Change-Id: I1ddcd752dd7034a9e121a9c99176929b92f2d24f
Bug-Url: https://bugzilla.redhat.com/1731395
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>