engine: Replace floppy file support for Ver >= 4.4
The sysprep floppy driver for windows is being replaced.

For versions 4.4 and greater we shall replace the floppy
that contains the sysprep functionality with an additional
CD ROM. We are still supporting the floppy driver for
versions less than 4.4 when a payload is specified
or when a floppy device path is specified.

Change-Id: I5d054089311e39f43db0c6c72873e22af94ce93d
Bug-Url: https://bugzilla.redhat.com/1306586
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>