engine: cpu flags handling in update cluster
It was not possible to edit cluster properties if cluster contained
host(s) not compatible with the configured CPU. It was because the CPU flags
were automatically updated to the configured values.

Now the CPU flags are not automatically updated, only if the
compatibility version or CPU name were changed.

Change-Id: If1a5207ae51e6ad50a905d1130b0d474dbf30afc
Bug-Url: https://bugzilla.redhat.com/1731395
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>