engine: ansible command race condition fix
Previously logFile used to be initialize in build() method.
Unfortunately that action was triggered by AnsibleExecutor
in async mode.
Multiple command used to rely on builder.logFile() and
called it just after executor attempted to run commands.
Due to slow nature of ansible command invocation the issue
did not manifest itself in production env.
However, for unit tests with mocks this was real issue.

This patch proposes to pass logFile reference in return value.

This patch is a prerequisite to linked bugzilla ticket.
More refactoring patches will follow in order to simplify and clarify
code.

Change-Id: I4df6b5b27328e5a3457a42ea2b9776ead205db3c
Bug-Url: https://bugzilla.redhat.com/1707451
Signed-off-by: Artur Socha <asocha@redhat.com>