engine: ansible command factory
This patch simlifies how ansible commands are setup by splitting data
setup and actual ansible command building. As a result
AnsibleCommandConfig was introduced and is supposed not to be modified
out of its creation place.

In the following patch several interface will be introduced so that all
consumer services worked only on read-only views of this configuration

Change-Id: I88a0da05db059a2ab24530a3891e8391351b29d5
Bug-Url: https://bugzilla.redhat.com/1707451
Signed-off-by: Artur Socha <asocha@redhat.com>