core: introduce CloneVmNoCollapse
CloneVmNoCollapse adds support to clone a VM
and keep its snapshots. This works by mapping the
original VM's snapshots to new IDs and copying
each snapshot to the new destination.
Later we create the snapshot along with its
original configuration (apart from network devices and disk
locations) in the database to properly support
preview and restore.

Change-Id: Ia98f20cb688b6af2f9f5e7c5042a7d3acd6c2316
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>