core: prevent teardown of a snapshot disk that is used by running VM
After disk image transfer ends, a teardown takes place for the image.
In case of snapshot disk that attached to a running VM, the teardown
will fail since the LV is still used by a VM and the snapshot disk
will remain locked.

This patch prevents from that teardown to happen.

Change-Id: Idfd120017841611b8ac9f65ee9b34d1134ca09cc
Bug-Url: https://bugzilla.redhat.com/1753168
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>