restapi: Allow editing disks using the REST API
In order to edit disk's parameters, the following request should be
sent on REST-API:
PUT /ovirt-engine/api/disks/123.

Currently, the only parameter that can be changed is the disk's
qcow version (which is relevant for VM disks only).

This patch changes BackendDiskResource so now it is possible to edit
disks using the request above.

Change-Id: I4d5a57e83a0d46e905a8bd6384483700f1150c5b
Bug-Url: https://bugzilla.redhat.com/1080097
Bug-Url: https://bugzilla.redhat.com/1652565
Bug-Url: https://bugzilla.redhat.com/854932
Signed-off-by: Shani Leviim <sleviim@redhat.com>