Update-Vm Unsupported Cluster Version
When updating a VM using rest-api and setting a custom
compatibility version which is higher that the version
of the Engine, an NPE is thrown in the Engine and the API
displays: 500: Internal Server Error.

This patch fixes this by validating that the provided
version is valid. The validation is done in the API layer.

Bug-Url: http://bugzilla.redhat.com/1724632
Change-Id: Id15fb491a870470bd93668e840e0ddcd76f4d32d
Signed-off-by: Ori_Liel <oliel@redhat.com>