ansible: Put host names into SAN in certificates
We currently put Hostname/IP into Common Name (CN) in the host
certificate subject.  This is not good for two reasons:

- Using CN for host name matching in certificates is obsolete and
  should be no longer used [1].

- CN may not contain an IP address.

In the latter case, migrations don't work on el8 since libvirt refuses
to connect to a destination host having invalid data in its
certificate.

The correct way to handle the certificates is to put the host name or
IP addresses to the Subject Alternative Name [2].

This patch fixes following flows:

1. Whe adding a new host, reinstalling existing host or enroling
   certificate we will always populate SAN part of host certificate.

2. When we check for certificate validity during host upgrade, we will
   renew the certificate not only when it's expired or going to expire
   soon, but also when host certificate doesn't have SAN field
   populated.

[1] http://wiki.cacert.org/FAQ/subjectAltName
[2] https://libvirt.org/remote.html#Remote_TLS_server_certificates

Change-Id: Ibadefde8cc0d58a858e17c971d1cb49026ba2a0a
Bug-Url: https://bugzilla.redhat.com/1763109
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>