core: Add default timezone in import vm
This patch will check the imported VM.
If the VM contains the value of TimeZone
"Hardware Clock Time Offset" it will be set
as is. In case it doesn't exist, it will be
set with the default engine's config.

For the guest agent status, it will check now
the VM vm_dynamic table, with reported values
and not consider the vm_static table that
is not reported by the agent of the VM.

Change-Id: Ied1294b81d7f14427e87267c3d8d857ef6999c11
Bug-Url: https://bugzilla.redhat.com/1743427
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>