webadmin: create list model with cluster default
Created a special entity model for lists that contain null as a
cluster default value. The null is added automatically to the first
position of the list in setItems method.

The model contains also the cluster value and sends cluster value
changed event if the value is changed. This allows the UI to listen
to such event and to be rendered correctly.

Change-Id: Ibccdef2cb84b4e2ec882ff30a141d9b44c40f369
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>