core: generate vnic alias if missing on hot unplug
Before [1] a vnic alias was generated by libvirt on the host and
reported to engine which would save it to the database. After [1] the
alias is generated by engine on VM creation using the device id and sent
to the host. It is later saved to the database when VM monitoring
updates engine with VM configuration. But during hot unplug, the alias
from the database is always used in case a VM which has been running
before upgrading engine to [1] has an old libvirt alias. Unfortunately
this created a loophole where a new engine-generated alias might not be
in the database (due to delayed VM monitoring) when hot unplug is
executed. This fails the host unplug action on the host due to the
missing alias. This fix generates the alias on the fly as in [1] during
hot unplug if it does not already exist. This should maintain backward
compatibility with old VMs and address the loophole.

[1] https://gerrit.ovirt.org/#/c/88149

Change-Id: Ia3621af8fb34863707852db73370fba7e9e890c5
Bug-Url: https://bugzilla.redhat.com/1766666
Signed-off-by: Eitan Raviv <eraviv@redhat.com>