core: Write disks under vdsm user when importing OVA
If VM disks are located on a NFS storage domain and the NFS directory is
exported with root_squash option, the disks are not writable by
processes running under root. In particular, this causes OVA import to
fail, because Ansible scripts used to read OVA are run under root.

This patch changes the scripts so as the part of the script that
accesses the disks is executed under vdsm user.

Change-Id: Ib01c987da4409aeef60ff2008b7a73c07fe52707
Bug-Url: https://bugzilla.redhat.com/1743543
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>