core: clear domains cache when changing state
The patch resolves an issue where the scheduling
filter (VmLeasesReadyFilterPolicyUnit) would use
stale data, leading to a host that is yet to acquire
the lease would attempt to run an HA VM.

This is caused by a race where the call to
GetStatsAsyncVDSCommand would finish processing after
the call to VmLeasesReadyFilterPolicyUnit#filter.
Since VmLeasesReadyFilterPolicyUnit uses the
domains list from VdsManager, it is only updated
after GetStatsAsyncVDSCommand is processed.
However, if a host is rebooted, there is a timeframe
in which there GetStatsAsyncVDSCommand was not
processed yet.  Which means the domains list will
hold the stale data, which was available
before the reboot.

The steps to hit the issue are usually the following
1. A host with an HA VM is rebooted
2. The VM attempts to run on another host but fails
   due to the lease still being held
3. The rebooted host is then coming up, but isn't
   fully initialized, and a call to GetStatsAsyncVDSCommand
   is yet to be made (this can be likely if the domain monitor
   in Vdsm is stuck)
4. The VM scheduler might try to run the VM on the
   rebooted host, and the filter, due to the fact
   that GetStatsAsync was not processed yet, will
   use the data from before step 1, in which the
   lease is reported as acquired
5. The VM will fail to run with a "No space left on device"
   as it has not acquired the lease yet

Change-Id: I8356912c50a900716bee3d50fe1cff654e80e696
 Bug-Url: https://bugzilla.redhat.com/1741625
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>