webadmin: vm popup - host cpu moved up in tab
On the VM popup dialog, Host tab, the CPU Passthrough was in the migration
section. Now it has its own section since it is not only related to the
migration and is also dependent on the selection of specific hosts.

Also fixed a small bug regarding enabling/disabling of CPU Passthrough
field if the VM type is High Performance.

Change-Id: I3ab8c0f1996d0b43647af753b50ed512a325743b
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>