webadmin: VmHostDevice action panel refactoring
It was not possible to inject action button on the detail sub tab page using
ui-extensions because the action panel was injected directly into
the view and was not processed by the presenter.

Now the action panel is injected into the presenter, the action buttons
from the extensions are added and it is bound to the sub tab view.

Change-Id: I8c2f6219d19b33bb20271669b6aaf759deb657d2
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>