licensing: automation: check-patch: Reject notices
Require new patches to include SPDX tags instead of the old
copyright/license notice.

It's enough to include the following two lines, in every file that
should be copyrighted:

Copyright oVirt Authors
SPDX-License-Identifier: Apache-2.0

The last one should be included (perhaps with a different license, if
needed) also if for any reason a copyright notice is still needed in the
file, and also if the copyright line should be different. E.g. if a file
is copied from some other project, with its own copyright and license.

Change-Id: I97720ec5dd7b39ca8028d3618aa2cf0608f4e387
Signed-off-by: Yedidyah Bar David <didi@redhat.com>