webadmin: fix same name template selection
This fix will let the user select a template in the webadmin
with the same name.

Previously, the template was chosen by the template name.
In case we have more then one data center, it is possible to create
template with the same name, it caused to select always the same
template.
With this fix, the template will be chosen by the template Guid.

Change-Id: I6c586a55d68df07c55500baa95efb691e54f7c98
Bug-Url: https://bugzilla.redhat.com/1590911
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>