core: Move fipsEnabled flag to VdsDynamic
The engine tries to read FIPS status from the host, as well as give
users ability to enable/disable FIPS. This was done using a single flag
in kernel line in VdsStatic.

This patch splits it into two distinct fields. Kernel command line in
VdsStatic keeps its current role: it's editable by users and is used to
enable/disable FIPS on host restart.

A new field in VdsDynamic contains FIPS status read from the host. It is
displayed in the UI, and it's mostly used by engine when a decision on
FIPS-specific functionality needs to be taken.

Change-Id: I051f3a8ca469a3aea66def24bb03d8e54260ce35
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1712325
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>