core: set new provider id on 'init' in AddProvider
Previously, the uid of a new provider to be added was set in
the 'execute' method of AddProviderCommand. That was confusing
because the provider is passed to the ProviderProxy in the
'validate' phase of the command already and yet when 'onAddition'
method of the ProviderProxy is called, the provided contains its
new uid. That's because we reference the provider that resides in
the parameters in both the command and the ProviderProxy.

In order to make it more clear and to avoid potential bugs when
the ProxyProvider clones the provider it gets, this patch changes
the place the uid is set to the 'init' method of the command.

Change-Id: Ie79b1e4dd3ace40bcfce74feab4e07bbb94cdfae
Signed-off-by: Arik Hadas <ahadas@redhat.com>