core: add boot uuid partition to cmdline
This patch will add in addition to the FIPS mode
of the host, the boot UUID of the partition.

Previously, the user would need to add the boot
partition manually in addition to the fips=1 flag
in the host kernel cmdline.
The value is taken from the host caps and set in vm_dynamic.
The user is still responsible to set the partition up.
In case the boot UUID already exists, it will be ignored
when setting the FIPS.

Change-Id: I04f9a097d1b5b470f998ae5c1de85982d2d0a2d1
Bug-Url: https://bugzilla.redhat.com/1692709
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>