webadmin: fetch 'in-preview' snapshot when commit
If a snapshot is in preview, and the page
refreshes, committing the snapshot will
raise an error, because of the
getSelectedItem() returns null.

In this patch, committing an in-preview
snapshot will work after refreshing
the page, because of every call to commit
fetches the IN_PREVIEW snapshot.

Screenshots: https://imgur.com/a/kCI76h3

Change-Id: Ib071083ccfb4825254fb337a30cfaa28bd0f6fcc
Bug-Url: https://bugzilla.redhat.com/1743690
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>