webadmin: Confirmation before SD format upgrade
screenshot: https://imgur.com/a/guSvhd2

Currently, when importing/attaching a storage domain to a DC with a
newer version than the storage format, the storage domain is being
upgraded automatically.
That way, it can't be attached back to an older DC version.

This patch adds a confirmation window before the upgrade occurs.
Pressing 'Cancel' cancels the operation.

Change-Id: I997e076c199d902f62c033b064d7af7e85c5968f
Bug-Url: https://bugzilla.redhat.com/1733031
Signed-off-by: Shani Leviim <sleviim@redhat.com>