core: prevent NPE on VMs monitoring
The method PollVmStatsRefresher::saveLastVmsList was not prepared for
a valid (yet rare) situation in which a pair of VM from the database
and a reported VM contains a VM that was running on the hosts in the
database and not being reported anymore by the host. In this case,
the VdsmVm would be null. The fix is to check whether VdsmVm is null
rather than VdsmVm::getVmDynamic that is never null.

Change-Id: I475a45637412a41957e4f0e5cb5e291de02e70d5
Bug-Url: https://bugzilla.redhat.com/1768174
Signed-off-by: Arik Hadas <ahadas@redhat.com>