vdsbroker: attestation service cleanup&refactoring
This is patch is a followup for getting rid of apache commons http client
in favour to httpcomponents. It provides changes that are supposed to
make the code more idiomatic and in addition several cleanups are
provided ie. visibility of methods/fields is adjusted. Obsolete 'static'
are removed where applicable.
No core logic should be changed.

Change-Id: Ie7c62e4a9bb6a224397f65241a676765359c2ab6
Signed-off-by: Artur Socha <asocha@redhat.com>