restapi: set portal group tag for iscsi storage connection
When sending a POST request to storage connections
endpoint with iscsi storage connection details
that includes 'portal' tag, the value of 'portal'
tag will be always set to '1' in the database -
(storage_server_connections) table.

Now, the value of 'portal' tag in the request
will be saved to the database.

Change-Id: Id7124811ac11870c20b2625cb8e457bf87320d85
Bug-Url: https://bugzilla.redhat.com/1768707
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>