core: EVENT_ID for rebooting a VM internally
When rebooting the VM from the OS level, no
EVENT_ID message was being logged so that
systems such as Cloudforms or ManageIQ could
could detect from the logs that the VM was
being rebooted.

Added EVENT_ID messages for rebooting that
are printed when the VM is rebooted at the
OS level. Now external systems can detect
when a VM is being rebooted.

Change-Id: I995cc99c37f9a63da679f6c5189f3086742413cf
Bug-Url: https://bugzilla.redhat.com/1710491
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>