core: Preserves Snapshot VmInit
Cloning a VM from a snapshot did not preserve
the VM Imitialization structure for Cloud Init
and Sysprep which meant that cloning a VM from
a snapshot did not provide the VM Init values
from the original VM giving incorrect settings.

With these changes the original VMs Cloud Init
or Sysprep settings are preserved by the snapshot
and frozen at the values set when the snapshot
was created so that cloning a VM from the snapshot
preserves the Cloud Init or Sysprep settings
accordingly.

Change-Id: Iac9c6178f02b100b2a0b2c8bf3f0da2ad5e0ceb2
Bug-Url: https://bugzilla.redhat.com/1739377
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>