webadmin+core: add RHCOS to operation systems
This patch introduces RHCOS to oVirt.
A new OS family is set, CoreOS. It integrates with the Ignition RFE.
When selecting a VM OS that is set as CoreOS, the WEB-UI will show
under the initial tab the option to set ignition.
More information about ignition in oVirt can be found at:
https://www.ovirt.org/develop/release-management/features/virt/coreos-ignition-support.html

The ignition option will be shown in the webadmin and in RunOnce.
Taking the arguments in the UI and building the ignition Json.
Note that the custom script area set as 'stronger' which means,
if a hostname is set twice it will take the custom script configuration.
As well for same user name configuration.

Change-Id: I3359e474eaa446c5b003199d72a6e61001108b76
Bug-Url: https://bugzilla.redhat.com/1726907
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>