webadmin: Same Name on Different VM Handling
Choosing Templates or VMs with multiple instances of the
same name on different Data Centers within the Template or
VM lists provided incorrect details for all but one of the
entries with the same name.

The issue of fetching the wrong VM details for Templates was
resolved on Gerrit Issue # 104303 while this change
addresses the VM List which also used the name to fetch the
VM details in the VM details screen.

Now we use the VM Id to choose the proper VM details.

This issue also addresses validating the VM name at the
front end when cloning a VM from a snapshot.

Now the validation for name uniqueness occurs at the front
end for consistency and correctness.

Change-Id: I22259b3740b0f80716d4dc1ade2de1aa96323755
Bug-Url: https://bugzilla.redhat.com/1743296
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>