engine: Created MDevTypes entity
Previously the only information stored about the mdevType was its name.

This commit introduces new entity type MDevType that allows to store
also the availableInstances and description of the mDev type. This
affected the HostDevice (change from Set<String> to List<MDevType>) and
also the format in which the mDevType data are stored in the database.

Change-Id: Ic9684aa22e4f1db923b53605a5e57684a1a02354
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>