core: EVENT_ID for rebooting a VM internally
When rebooting the VM from the OS level, no
EVENT_ID message was being logged so that
systems such as Cloudforms or ManageIQ could
could detect from the logs that the VM was
being rebooted.

Added EVENT_ID messages for rebooting that
are printed when the VM is rebooted at the
OS level. Now external systems can detect
when a VM is being rebooted.

Change-Id: Ib7cc15c6b7e8f5b50d831540abc5268a13057afc
Bug-Url: https://bugzilla.redhat.com/1710491
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>