core: Add qemu:capabilities to the VM XML
The new qemu:capabilities is needed in order to support experimental
incremental backup feature, it should be removed when the feature will
be fully supported by libvirt in downstream version.

This section should be added only if the seleceted host supports
incremental backup feature.

The following element should be added:

<qemu:capabilities>
    <qemu:add capability='blockdev'/>
    <qemu:add capability='incremental-backup'/>
</qemu:capabilities>

https://www.redhat.com/archives/libvir-list/2019-November/msg01315.html

Change-Id: I3aba3c28fcb5cc88a9eea5ac2e6dd45a9ee0d959
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>